add two parameters for AgentTicketCloseBulk
To be able to redirect to ticket zoom by using some parameters, the parameters 'OrigTicketID' and 'GoToZoom' were added.
Now it is possible to create a link anywhere in OTRS and force a redirect to the ticket zoom of 'OrigTicketID'. Until now
the user was redirected to the last overview or - if the TicketZoom attribute was set in the quickclose data - to the ticket.